+++ test/index.js
@@ -10,21 +10,43 @@ describe('DoubleClick Floodlight', function(){
   var settings;
   var floodlight;
   var test;
+  var baseEndpoint = 'https://ad.doubleclick.net/ddm/activity/';
 
   beforeEach(function(){
     settings = {
       source: '654757884637545',
-      events: [{
-        key: 'Application Installed',
-        value: {
-          cat: 'activityTag',
-          type: 'groupTag',
-          customVariable: [{
-            key: 'version',
-            value: 'u1'
-          }]
+      events: [
+        {
+          key: 'Application Installed',
+          value: {
+            event: 'Application Installed',
+            cat: 'activityTag',
+            type: 'groupTag',
+            customVariable: [{
+              key: 'version',
+              value: 'u1'
+            }]
+          }
+        },
+        {
+          key: 'Rick and Morty Season 3 is the only hope we have',
+          value: {
+            event: 'Rick and Morty Season 3 is the only hope we have',
+            cat: 'wubbalubbadubdub',
+            type: 'noType',
+            customVariable: []
+          }
+        },
+        {
+          key: 'Rick and Morty Season 3 is the only hope we have',
+          value: {
+            event: 'Rick and Morty Season 3 is the only hope we have',
+            cat: 'showmewhatyougottttt',
+            type: 'noType',
+            customVariable: []
+          }
         }
-      }]
+      ]
     };
   });
 
@@ -74,11 +96,16 @@ describe('DoubleClick Floodlight', function(){
     });
   });
 
-   describe('mapper', function(){
+  describe('mapper', function(){
+    before(function(){
+      // stubbing the random cachebuster for testing
+      // this gets multiplied by 10000000000000000000 so will hardcode ord
+      sinon.stub(Math, 'random').returns(0.27005030284556764);
+    });
+
     describe('track', function(){
       it('should map application installed', function(){
-        // ignoring `ord` since it is randomly generated
-        test.maps('app-install', settings, { ignored: 'ord' });
+        test.maps('app-install', settings);
       });
 
       it('should not send unmapped events', function(){
@@ -90,24 +117,12 @@ describe('DoubleClick Floodlight', function(){
   describe('track', function(){
     it('should track application installed correctly', function(done){
       var json = test.fixture('app-install');
-      var output = json.output;
-      // stubbing the random cachebuster for testing
-      // this gets multiplied by 10000000000000000000 so will hardcode ord
-      sinon.stub(Math, 'random').returns(0.27005030284556764);
-
-      var expectedUrl = 'https://ad.doubleclick.net/ddm/activity/'
-        + 'dc_rdid=' + output.dc_rdid
-        + ';src=' + output.src
-        + ';cat=' + output.cat
-        + ';type=' + output.type
-        + ';ord=' + '2700503028455676400'
-        + ';tag_for_child_directed_treatment=' + output.tag_for_child_directed_treatment
-        + ';dc_lat=' + output.dc_lat;
+      var expectedUrl = baseEndpoint + json.output.endpoints[0];
 
       test
         .track(json.input)
         .expects(200)
-        .set('User-Agent', output.userAgent)
+        .set('User-Agent', json.output.userAgent)
         .end(function(err, res){
           if (err) throw err;
           // we don't use query params so making sure the endpoint
@@ -117,8 +132,26 @@ describe('DoubleClick Floodlight', function(){
         });
     });
 
-    it('should set copa compliance', function(done) {
-      var json = test.fixture('app-install');
+    it('should send properties as custom variables', function(done){
+      var json = test.fixture('custom-track');
+      var expectedUrl = baseEndpoint + json.output.endpoints[0];
+
+      test
+        .track(json.input)
+        .expects(200)
+        .set('User-Agent', json.output.userAgent)
+        .end(function(err, res){
+          if (err) throw err;
+          // we don't use query params so making sure the endpoint
+          // which includes the payload is correct
+          assert.strictEqual(res[0].request.url, expectedUrl);
+          done();
+        });
+    });
+
+    it('should set copa compliance', function(done){
+      var json = test.fixture('app-install-copa');
+      var expectedUrl = baseEndpoint + json.output.endpoints[0];
 
       // set integration option
       json.input.integrations = {
@@ -126,25 +159,55 @@ describe('DoubleClick Floodlight', function(){
           copaCompliant: true
         }
       };
-      var output = json.output;
-      var expectedUrl = 'https://ad.doubleclick.net/ddm/activity/'
-        + 'dc_rdid=' + output.dc_rdid
-        + ';src=' + output.src
-        + ';cat=' + output.cat
-        + ';type=' + output.type
-        + ';ord=' + '2700503028455676400'
-        + ';tag_for_child_directed_treatment=1'
-        + ';dc_lat=' + output.dc_lat;
 
       test
         .track(json.input)
         .expects(200)
-        .set('User-Agent', output.userAgent)
+        .set('User-Agent', json.output.userAgent)
         .end(function(err, res){
           if (err) throw err;
           assert.strictEqual(res[0].request.url, expectedUrl);
           done();
         });
+    });
+
+    it('should send multiple requests if multiple tags are mapped for an event', function(){
+      var json = test.fixture('multiple-track');
+      var expectedUrl1 = baseEndpoint + json.output.endpoints[0];
+      var expectedUrl2 = baseEndpoint + json.output.endpoints[1];
+
+      test
+        .track(json.input)
+        .set('User-Agent', json.output.userAgent)
+        .requests(2);
+
+      test
+        .request(0)
+        .expects(200)
+        .end(function(err, res){
+          if (err) throw err;
+          assert.strictequal(res[0].request.url, expectedurl1);
+          done();
+        });
+
+      test
+        .request(1)
+        .expects(200)
+        .end(function(err, res){
+          if (err) throw err;
+          assert.strictequal(res[0].request.url, expectedurl2);
+          done();
+        });
+    });
+
+    it('should not send any requests for unmapped events', function(done){
+      var json = test.fixture('app-install');
+      json.input.event = 'I love Kanye';
+
+      test
+        .track(json.input)
+        .requests(0)
+        .end(done);
     });
   });
 });