+++ HISTORY.md
@@ -1,10 +1,15 @@
 
+1.0.2 / 2016-11-09
+==================
+
+  * Fix bug with mapping custom variables
+
 1.0.1 / 2016-08-17
 ==================
 
-  * update validation logic for mapped events
+  * Update validation logic for mapped events
 
 1.0.0 / 2016-08-11
 ==================
 
-  * initial release 
+  * Initial release